refactor(saga): split schema.go into schema, registry, and proto_resolve
Split the 853-line schema.go into three files:
- schema.go (415 lines): types, constants, errors, validation
- registry.go (280 lines): Registry type and methods
- proto_resolve.go (168 lines): proto type resolution

Remove schema.go from knownOversizedFiles since all files are now under 800 lines.
