+++ shared/pkg/saga/schema/schema.go
@@ -1,20 +1,12 @@
 // Package schema provides YAML-based handler schema definitions for saga orchestration.
 // It enables compile-time validation and IDE support for handler references in Starlark scripts.
-//
-//meridian:large-file - known oversized file; split tracked in backlog
 package schema
 
 import (
 	"errors"
 	"fmt"
-	"os"
-	"path/filepath"
-	"sort"
 	"strings"
-	"sync"
 
-	"google.golang.org/protobuf/reflect/protoreflect"
-	"google.golang.org/protobuf/reflect/protoregistry"
 	"gopkg.in/yaml.v3"
 )
 
@@ -420,434 +412,4 @@ func (f *FieldDef) validateEnumValue(name string, params map[string]any) error {
 		}
 	}
 	return fmt.Errorf("%w: %s got %q, allowed: %v", ErrInvalidEnumValue, name, strVal, f.Values)
-}
-
-// DeprecatedMapping records how a deprecated handler name maps to its current replacement.
-type DeprecatedMapping struct {
-	// CurrentName is the fully-qualified name of the current handler.
-	CurrentName string
-
-	// ConversionRule is the conversion rule that applies.
-	ConversionRule *ConversionRule
-}
-
-// Registry manages multiple handler schemas.
-type Registry struct {
-	mu              sync.RWMutex
-	schemas         []*Schema
-	handlers        map[string]*HandlerDef
-	deprecatedNames map[string]*DeprecatedMapping // old name -> current handler mapping
-}
-
-// NewRegistry creates a new empty schema registry.
-func NewRegistry() *Registry {
-	return &Registry{
-		schemas:         make([]*Schema, 0),
-		handlers:        make(map[string]*HandlerDef),
-		deprecatedNames: make(map[string]*DeprecatedMapping),
-	}
-}
-
-// LoadFromYAML parses and loads a schema from YAML bytes.
-func (r *Registry) LoadFromYAML(data []byte) error {
-	schema, err := Parse(data)
-	if err != nil {
-		return err
-	}
-
-	r.mu.Lock()
-	defer r.mu.Unlock()
-
-	r.schemas = append(r.schemas, schema)
-	for name, handler := range schema.Handlers {
-		r.handlers[name] = handler
-		// Index deprecated name mappings from conversion rules
-		for i := range handler.Conversions {
-			conv := &handler.Conversions[i]
-			if conv.FromName != "" {
-				if existing, exists := r.deprecatedNames[conv.FromName]; exists {
-					return fmt.Errorf(
-						"%w: duplicate deprecated alias %q maps to both %s and %s",
-						ErrInvalidConversionRule, conv.FromName, existing.CurrentName, name,
-					)
-				}
-				r.deprecatedNames[conv.FromName] = &DeprecatedMapping{
-					CurrentName:    name,
-					ConversionRule: conv,
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
-// GetHandler returns the handler definition for the given name.
-func (r *Registry) GetHandler(name string) (*HandlerDef, error) {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	handler, ok := r.handlers[name]
-	if !ok {
-		return nil, fmt.Errorf("%w: %s", ErrHandlerNotFound, name)
-	}
-	return handler, nil
-}
-
-// ListHandlers returns a sorted list of all registered handler names.
-func (r *Registry) ListHandlers() []string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	names := make([]string, 0, len(r.handlers))
-	for name := range r.handlers {
-		names = append(names, name)
-	}
-	sort.Strings(names)
-	return names
-}
-
-// HasHandler returns true if the handler is registered.
-func (r *Registry) HasHandler(name string) bool {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	_, ok := r.handlers[name]
-	return ok
-}
-
-// LookupDeprecated checks if a handler name is deprecated and returns
-// the mapping to the current handler, or nil if the name is not deprecated.
-func (r *Registry) LookupDeprecated(name string) *DeprecatedMapping {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	return r.deprecatedNames[name]
-}
-
-// IsDeprecated returns true if the handler exists but is marked as deprecated.
-func (r *Registry) IsDeprecated(name string) bool {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	handler, ok := r.handlers[name]
-	if !ok {
-		return false
-	}
-	return handler.Deprecated
-}
-
-// LoadFromFile loads a schema from a YAML file.
-func (r *Registry) LoadFromFile(path string) error {
-	data, err := os.ReadFile(path)
-	if err != nil {
-		return fmt.Errorf("failed to read schema file %s: %w", path, err)
-	}
-	return r.LoadFromYAML(data)
-}
-
-// ListSchemas returns all loaded schemas in the registry.
-func (r *Registry) ListSchemas() []*Schema {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	result := make([]*Schema, len(r.schemas))
-	copy(result, r.schemas)
-	return result
-}
-
-// LoadFromDirectory loads all YAML schema files from a directory.
-// Files must have .yaml or .yml extension. Subdirectories are not traversed.
-func (r *Registry) LoadFromDirectory(dir string) error {
-	entries, err := os.ReadDir(dir)
-	if err != nil {
-		return fmt.Errorf("failed to read schema directory %s: %w", dir, err)
-	}
-
-	for _, entry := range entries {
-		if entry.IsDir() {
-			continue
-		}
-
-		ext := filepath.Ext(entry.Name())
-		if ext != ".yaml" && ext != ".yml" {
-			continue
-		}
-
-		path := filepath.Join(dir, entry.Name())
-		if err := r.LoadFromFile(path); err != nil {
-			return fmt.Errorf("failed to load schema %s: %w", path, err)
-		}
-	}
-
-	return nil
-}
-
-// ToSchema returns a Schema snapshot from the registry's current state.
-// This enables callers that have a YAML-based Registry to use BuildServiceModulesFromSchema.
-func (r *Registry) ToSchema() *Schema {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	handlers := make(map[string]*HandlerDef, len(r.handlers))
-	for name, def := range r.handlers {
-		handlers[name] = def
-	}
-	return &Schema{Handlers: handlers}
-}
-
-// NewRegistryFromSchema creates a Registry pre-populated with handler definitions from a Schema.
-// This is the inverse of ToSchema() and enables creating a Registry from proto-derived schemas.
-// It also rebuilds the deprecatedNames index from conversion rules that specify FromName.
-func NewRegistryFromSchema(s *Schema) *Registry {
-	r := NewRegistry()
-	if s != nil {
-		for name, def := range s.Handlers {
-			r.handlers[name] = def
-			// Rebuild deprecated name mappings from conversion rules
-			for i := range def.Conversions {
-				conv := &def.Conversions[i]
-				if conv.FromName != "" {
-					r.deprecatedNames[conv.FromName] = &DeprecatedMapping{
-						CurrentName:    name,
-						ConversionRule: conv,
-					}
-				}
-			}
-		}
-	}
-	return r
-}
-
-// ValidateHandlerParams validates parameters for a named handler.
-// Returns ErrHandlerNotFound if the handler schema is not registered.
-func (r *Registry) ValidateHandlerParams(handlerName string, params map[string]any) error {
-	handler, err := r.GetHandler(handlerName)
-	if err != nil {
-		return err
-	}
-	return handler.ValidateParams(params)
-}
-
-// LinterMetadata describes handler characteristics needed by the semantic linter.
-type LinterMetadata struct {
-	// IsExternal indicates the handler calls external systems (non-idempotent).
-	IsExternal bool
-
-	// RequiresPreCheck indicates verify_external_state must be called before this handler.
-	RequiresPreCheck bool
-
-	// CompensationStrategy indicates how compensation is handled ("auto", "saga_managed", "none", or "").
-	CompensationStrategy string
-
-	// HasAutoCompensation indicates the handler has a compensate: field.
-	HasAutoCompensation bool
-
-	// IsDeprecated indicates the handler is deprecated and should produce a warning.
-	IsDeprecated bool
-
-	// ReplacedBy is the current handler name if this handler has been superseded.
-	ReplacedBy string
-}
-
-// BuildLinterMetadata extracts linter metadata from the schema registry.
-// Returns a map of handler names to their metadata for linter validation.
-// All handlers are included to support compensation coverage checks.
-func (r *Registry) BuildLinterMetadata() map[string]LinterMetadata {
-	metadata := make(map[string]LinterMetadata)
-
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	for name, handler := range r.handlers {
-		meta := LinterMetadata{}
-		if handler.External {
-			meta.IsExternal = true
-			meta.RequiresPreCheck = true
-		}
-		if handler.Compensate != "" {
-			meta.HasAutoCompensation = true
-			meta.CompensationStrategy = string(CompensationStrategyAuto)
-		} else {
-			meta.CompensationStrategy = string(handler.CompensationStrategy)
-		}
-		if handler.Deprecated {
-			meta.IsDeprecated = true
-		}
-		metadata[name] = meta
-	}
-
-	// Also add entries for deprecated name aliases so they can be detected
-	for oldName, mapping := range r.deprecatedNames {
-		if _, exists := metadata[oldName]; !exists {
-			// Get the current handler's metadata as a base
-			if currentMeta, ok := metadata[mapping.CurrentName]; ok {
-				deprecatedMeta := currentMeta
-				deprecatedMeta.IsDeprecated = true
-				deprecatedMeta.ReplacedBy = mapping.CurrentName
-				metadata[oldName] = deprecatedMeta
-			}
-		}
-	}
-
-	return metadata
-}
-
-// ResolveProtoTypes resolves proto-referenced handlers in the schema by looking up
-// proto service/method descriptors and populating Params/Returns from proto reflection.
-// Handlers without ProtoRef are left unchanged (legacy inline format).
-// Uses the global proto registry by default; pass a custom resolver for testing.
-func (s *Schema) ResolveProtoTypes(files *protoregistry.Files) error {
-	if files == nil {
-		files = protoregistry.GlobalFiles
-	}
-	for handlerName, handler := range s.Handlers {
-		if handler.ProtoRef == nil {
-			continue
-		}
-		if err := resolveHandlerProto(handlerName, handler, files); err != nil {
-			return fmt.Errorf("handler %s: %w", handlerName, err)
-		}
-	}
-	return nil
-}
-
-// resolveHandlerProto resolves a single handler's proto reference into Params/Returns.
-func resolveHandlerProto(_ string, handler *HandlerDef, files *protoregistry.Files) error {
-	ref := handler.ProtoRef
-
-	// Parse "package.Service/Method" into service and method names
-	parts := strings.SplitN(ref.FullMethod, "/", 2)
-	if len(parts) != 2 {
-		return fmt.Errorf("%w: %s", ErrInvalidProtoRPC, ref.FullMethod)
-	}
-	serviceFQN := protoreflect.FullName(parts[0])
-	methodName := protoreflect.Name(parts[1])
-
-	// Look up the service descriptor
-	serviceDesc, err := findServiceDescriptor(files, serviceFQN)
-	if err != nil {
-		return err
-	}
-
-	// Look up the method
-	methodDesc := serviceDesc.Methods().ByName(methodName)
-	if methodDesc == nil {
-		return fmt.Errorf("%w: %s in service %s", ErrProtoMethodNotFound, methodName, serviceFQN)
-	}
-
-	// Resolve params from request message
-	reqMsg := methodDesc.Input()
-	params, err := resolveExposedFields(reqMsg, ref.ExposedParams, ref.ParamAliases)
-	if err != nil {
-		return fmt.Errorf("params: %w", err)
-	}
-	handler.Params = params
-
-	// Resolve returns from response message
-	respMsg := methodDesc.Output()
-	returns, err := resolveExposedFields(respMsg, ref.ExposedReturns, nil)
-	if err != nil {
-		return fmt.Errorf("returns: %w", err)
-	}
-	handler.Returns = returns
-
-	return nil
-}
-
-// findServiceDescriptor searches the proto registry for a service by fully-qualified name.
-func findServiceDescriptor(files *protoregistry.Files, fqn protoreflect.FullName) (protoreflect.ServiceDescriptor, error) {
-	desc, err := files.FindDescriptorByName(fqn)
-	if err != nil {
-		return nil, fmt.Errorf("%w: %s (%w)", ErrProtoServiceNotFound, fqn, err)
-	}
-	sd, ok := desc.(protoreflect.ServiceDescriptor)
-	if !ok {
-		return nil, fmt.Errorf("%w: %s is not a service descriptor", ErrProtoServiceNotFound, fqn)
-	}
-	return sd, nil
-}
-
-// resolveExposedFields builds a FieldDef map from a proto message descriptor,
-// filtered to only the exposed field paths. If exposed is nil/empty, all top-level
-// fields are included. Aliases are applied to param fields.
-// Returns ErrProtoFieldPathNotFound if any exposed path cannot be resolved.
-func resolveExposedFields(md protoreflect.MessageDescriptor, exposed []string, aliases map[string]string) (map[string]*FieldDef, error) {
-	fields := make(map[string]*FieldDef)
-
-	if len(exposed) == 0 {
-		// Include all top-level fields
-		fds := md.Fields()
-		for i := 0; i < fds.Len(); i++ {
-			fd := fds.Get(i)
-			fieldName := string(fd.Name())
-			fields[fieldName] = deriveFieldDef(fd)
-		}
-	} else {
-		// Only include exposed field paths
-		for _, path := range exposed {
-			fd := resolveFieldPath(md, path)
-			if fd == nil {
-				return nil, fmt.Errorf("%w: %q in message %s", ErrProtoFieldPathNotFound, path, md.FullName())
-			}
-			// Use the leaf field name as the key
-			leafName := leafFieldName(path)
-			if _, dup := fields[leafName]; dup {
-				return nil, fmt.Errorf("%w: %q from path %q in message %s", ErrDuplicateLeafName, leafName, path, md.FullName())
-			}
-			fields[leafName] = deriveFieldDef(fd)
-		}
-	}
-
-	// Apply aliases with validation
-	for original, alias := range aliases {
-		def, ok := fields[original]
-		if !ok {
-			return nil, fmt.Errorf("%w: %q (alias target: %q)", ErrUnknownAliasSource, original, alias)
-		}
-		if _, collision := fields[alias]; collision && alias != original {
-			return nil, fmt.Errorf("%w: %q (from alias of %q)", ErrAliasCollision, alias, original)
-		}
-		delete(fields, original)
-		fields[alias] = def
-	}
-
-	return fields, nil
-}
-
-// resolveFieldPath resolves a dot-separated field path (e.g., "log.status_tracking.current_status")
-// through nested proto message descriptors. Returns the leaf field descriptor, or nil if not found.
-func resolveFieldPath(md protoreflect.MessageDescriptor, path string) protoreflect.FieldDescriptor {
-	parts := strings.Split(path, ".")
-	current := md
-
-	for i, part := range parts {
-		fd := current.Fields().ByName(protoreflect.Name(part))
-		if fd == nil {
-			return nil
-		}
-		// If this is the last part, return the field descriptor
-		if i == len(parts)-1 {
-			return fd
-		}
-		// Otherwise, navigate into the nested message
-		if fd.Kind() != protoreflect.MessageKind {
-			return nil // Can't traverse into non-message field
-		}
-		current = fd.Message()
-	}
-	return nil
-}
-
-// leafFieldName returns the last segment of a dot-separated path.
-func leafFieldName(path string) string {
-	if idx := strings.LastIndex(path, "."); idx >= 0 {
-		return path[idx+1:]
-	}
-	return path
-}
-
-// HasProtoRef returns true if the handler uses proto-referenced format.
-func (h *HandlerDef) HasProtoRef() bool {
-	return h.ProtoRef != nil
 }
